Add ipv4 address to mesh interfaces
This generates and adds an ipv4 address to mesh interfaces to allow
devices to be reached across ipv4 only antennas. This new address is
generated at boot time and stored in the config.

In the case of an ipv4 antenna it can manually be made to work by adding
the ipv4 addresses of devices on one end of the link to the manual ip list
of devices on the other end of the link.