Add router ip to light client tunnels
Previously phone clients could not properly send traffic becuase the
router had no ip on the wg tunnel with which to send a response. This
is now corrected.

The solution here highlights how the way light client traffic is handled
is nonstandard for openwrt. The 'openwrt way' to perform the task of natting
traffic from many interfaces is seen for the lan bridge. br-lan is a single
bridged interface representing many physical interfaces, the nat is then done
on this single virtual interface and likewise it is the only one assigned
the ipv4 lan ip.

While this is simplifying in a static usecase it's not so simplifying in a
dyanmic one, since these WireGuard interfaces for light clients are created
and removed in a dyanmic manner managing the bridge is simply another complication
that we can hopefully avoid by directly inserting a /32 ip for ourselves and a /32
route for the target device