Compression for usage history
this adds compression for the usage history file, old style files
should be read correctly and then compressed on the next write, this
results in a 85% reduction in filesize. 377KB/2.6MB in my test. Using
the highest compression ratio results in 322KB/2.6MB or a pretty minimal
improvement in exchange for a lot more computation time. Since this is
blocking the main thread it's probably not worth it.

The XZ algorithm would save another 70KB or so with a 250KB/2.6MB tested
size, but we would have to import more code (this will dedupe with
the compressed log crate) so I won't pursue that optimization