Protect against slow transaction verification
Over the Christmas - New Year break we experienced an issue with the
exit where some network disruption threw the Tokio futures executor
into a bad state. Any web requests and some local requests would take
more than a second to process. Despite this constant delay everything
continued to function fine.

Becuase the timeout for the transaction verification was only 1 second
the exit would never validate any payment sent to it, causing clients
to resend their payments every 15 minutes as their last attempt failed.

This went on indefinately and resulted in significant wallet balance draining.

Clients do estimate their own usage and it would be possible for them to compare
the desired bill to their sync value with the exit but exactly what to do about
such a conflict is questionable, the exit will start enforcing on them anyways
and there's nothing they can do to get the exit back into a working state.

The best resolution for this issue that I've been able to come up with is both
an emergency crash check that detects when the exit is out of sync and a dramatic
increase in the maximum allowed verification time for transactions on the exits.
This means that even if the exit is on fire and growing memory usage constantly
as futures can't terminate faster than they are issued it will at least accept
payments and quickly trigger it's own restart.