Allow manual peers for non-gateways
At this point we should probably seperate these into 'exit manual peers' and
actual manual peers. Right now devices ship with a series of domain names for
'manual peers' these are peers where normal peer discovery is not done automatically
but instead we reach out to them directly.

For exit addresses we use the various domain names of the exits in order to get gateways
to dail them up over the internet and pull them into the mesh. But manual peers may still
be used for general purpose addition of peers that won't peer for other reasons. This may
or may not involve gateways.

This change is sort of a hack which allows us to access manual peers that are pure ip addresses
even when the device is not a gateway. This is safe because no dns resolution is needed and
we don't ship devices with any ip's for manual peers. But like I said before this strange
edge case where manual peers work somtimes but not others probably means we should rename
the list of exits to 'exit peers' and append it to manual peers if we are a gateway