Expand dialer ip generation subnet
So we want devices to assign themselves routable ip's so that they can be contacted
remotely even over ipv4 only interfaces. The problem is that if we have them assign
themselves only a /32 they will not be able to respond to any device that contacts
them.

This patch solves that problem by having them generate within 192.168.100.0/18 which
can't possibly conflict with 192.168.10.0/24 (the router internal range) nor many of
the other internal ranges which go from 192.168.0-192.168.100. They then assign their
local subnet as a /24. That way devices attempting to contact them can give themselves
an ip on the same subnet to facilicate cross communication.

Overall this is pretty overcomplicated compared to SLACC and I expect it to either cause
some sort of problem eventually or be eliminated in favor of not using antennas without
any form of ipv6 support. But it's worth exploring for now