Revert ipv4 dialer address changes
This reverts a series of commits that added an ipv4 address to interfaces
on devices in the hopes that we would be able to manually address other
routers over bridges that did not support ipv6. Sadly this feature is a
real rabbit hole filled with gotcha's at every level.

We'll just have to avoid these antenna models