Clean up features usage
Now that we're doing some more conditional compilation both for the exit
in the form of longer timeouts and the client in the form of client specific
code in the common components it's best to clean up the feature list and make
them requirements for binary building.

For example while the light_client_support feature exists for rita_common it doesn't
really make sense to build rita_client without it. As the other components aren't designed
to be removed.