Light client billing
This adds billing for light client usage, which is actually really
simple. There are no destinations to care about and only our own
price to consider so we can just monitor the light client tunnel
grab it's usage counter and add the resulting amount to debt keeper.

I've pulled a lot of utility functions from the exit billing code since
it works on the same principle.