Remove light_client_enabled feature
In order to do this we move light client GC over to the light
client module, where it involves some extra iterating in order to find
the data we need rather than conditional compilation.

Ultimately the build tools complexity of the alternative was far too high
to really be worth the very minimum runtime impact. While a the time complexity
of this function is high the actual number of light client addresses is hard
limited at 256 and will often be much less or even zero.