Update client traffic watcher to compensate for WireGuard overhead
After investigating some billing inaccuracies on the exit I determined
that wireguard counters where counting significantly less traffic
than the relay traffic counters. This means that the amount being
paid to the exit is less than the amount paid to the relay.

This inaccuracy is due to the fact that there is 80 byters of overhead
per packet that was not being considered.

This is only the client side fix, which is much less complex than the exit
fix which will require a move of exit billing back to ipset counters
rather than wireguard.