Update interface toggling for new network layout
With the addition of light client support both the network and wireless
configs have changed pretty significantly.

This commit updates ethernet port toggling to work as expected given the
new config and removes the ability to toggle wireless networks as that
functionality needs a larger re-think and some new endpoints given
the changes made in /etc/config/wireless

also included in this patch is the ability to display StaticWAN with
all the configured info from the get /interfaces endpoint which I'm
sure will be a dashboard improvement.