Long timeouts by default on x86
this fixes the problem of always needing to specify the server feature
when building the exit software. It's also based on the assumption that
bigger x86 machines will be able to handle the memory pressure this may
cause during transient network events.

No system will be able to survive prolonged timeouts of course as that's
a memory leak by definition.