Fallback billing for client
This adds local billing as a 'fallback option' if requests to the exit
for exit debts fail.

This is pretty similar to what we do for gateways now but instead of
happening when we detect that we are a gateway we do send a debt update
any time requesting debts from the exits fails. If requests always fail
billing will be performed locally. If a request succeeds we will update
with the exit value again and continue on as if we had never sent a local
update