Drop old debt database entries
We maintain a persistant database of debts to handle the case where
a device is rebooted after slowly building up a bill with another user.

This patch does not save debts entires that have not been touched for
more than 24 hours.

This should save on disk wries and prevent pollution of the debts database.
The consideration to be made here is how this could cause errors. This could
cause errors by dropping debts to an upstream node that we haven't seen recently
but still owe some small amount of money if it remembered that we owed it money.

Worth considering as that's a new edge case.