We can't discard entries with incoming payments
In my pervious commit 01b33f8d42a7b55149191949432e9538d158e461 I was
incorrect about the impacts of the change. If a user where to have a credit in
incoming payments waiting to be applied the previous change would wipe it away
on exit reboot.

The previous patch was not yet deployed so thankfully this did not occur.