Apply incoming payment credit immediately
As outlined in the comment in this commit and in the previous
commit it is possible for routers to build up a significant incoming
payments balance to the exit.

In order to resolve this we should always attempt to apply this credit
during every billing cycle. I'm still considering if this will have
any unintended side effects. I don't believe it will at this point
but it requires more consideration.