Increase accuracy of dao payments
The DAO manager code is very straightforward and would mark the fee
as paid if it had made an attempt to send a transaction. No matter
how successful or unsuccessful that attempt was.

In the last week or so $30 in subnet dao fees where not collected
due to routers with low balances. $1 was not collected due to other
errors.

This change does not persist the last subnet dao payment, it's difficult
to serialize instants. But a router left on will now at least attempt
to settle up.