Simulated TxFees
This adds a configurable txfee that is by default sent to the company
behind the Althea name Hawk Networks INC. The idea here is that Peggy
issues currently prevent us from launching the Althea blockchain as
planned.

But we have been successful at starting networks and generally growing
the Althea ecosystem. We would like to get our fee in place early before
a migration to the Althea blockchain which will come at an undefined
point in the future.

With that in mind we've consulted with all of our online networks and
they have agreed that collecting the fee in a simulated manner is the
best path forward for everyone.

The default is currently set at 10% but this may be modified pending
future developments.