Don't discard what we owe people
Once again the change in 6093ad99bbc283d941d2e01801e62ca736b43613 was
flawed. Considering how many iterations of this patch have been flawed
this patch also adds some test cases to prevent this from happening
in the future.

To address the actual issue we would set all debts outside of the
special discard case to zero which was not good