Add /prices endpoint
The prices endpoint gets the actual prices you are currently paying to everyone
this is useful in that we provided good ways to get past paid amounts but
not present pricing.

It also moves the pricing file from common endpoints to client endpoints
since things like auto pricing are just forced disabled on the exits anyways
the accessiblity functions really where not being used.