Toggle endpoints for mesh and lightclient WiFi
This is a refactor of the interfaces behavior when it comes to wireless
endpoints.

Rather than having one for each wireless interface and genearlly a lot
of confusion it's been boiled down to a single lightclient and mesh network
per device that's then toggled off and on. This brings the number of possible
states into somthing manageable with a toggle switch on the dashboard.