Fix xdai gas computation
When we first moved to the xdai chain most of the focus was on more
reliable, not cheaper transactions. So I neglected to optimize the gas
consumption. It also wasn't exactly clear what the gas cost was at the
time as the gas price endpoint would return zero.

Now this seems to be out of the way and we can very dramatically reduce
the amount spent on gas and still keep things functional.