Handle the no results case properly
The 'more than one entry' case was actually handling the 'zero entry'
path.

This of course didn't work, so now we handle zero entries as well as
more than one.

I have otherwise changed the behavior of this function by inserting
a panic on duplicate entries. I think this is appropriate and hope
it won't cause problems.