WIP: Pull in fastping for parallel exit pinging
WARNING: runs out of file descriptors

This refactors the exit ping code such that it's in pure rust
and done in parallel rather than shelling out to the system ping

This makes the exit endpoint operate in somthing much closer to
constant time (barring scheduling shenenegans)

This biggest thing to really mention here is the addition of the fastping
create into our own repo. It didn't have a crates.io crate and I don't
care for another git dependency. We already have too many of those.