OracleUpdate modernization
This struct has been around for a while and is used pretty heavily in actual
day to day operations to drive updates.

But the existance of a system_chain selection is problematic.

Routers are on a default oracle xdaiprices or ethprices, these must
contain system_chain values to be read. But if you go and try to change
your system_chain yourself you create a race condition, if the oracle is
running in parallel the system chain will snap back to the assigned value.

Making these optional prevents that strange snap back behavior.

I also removed the fudge factor and payment multiplier, in the past
they had to be present to allow us to change those settings at all.
But now that we have the general purpose config merger they aren't
required and are so rarely changed that they cause more problems than
they solve.