Cache most exit vars
This patch caches all the exit settings referneced by the endpoints.

As exits scale more and more threads get thrown at their threadpools
resulting in rising and eventually untenable lock contention on the main
SETTINGS lock.

Most things behind that lock we never really want to change at runtime.
other things like the price we do.

This patch isn't the best soltuion to the problem by far, moving the locks down
to a more granular level within SETTINGS is the real solution here. But that would
have wide reaching syntax changes throughout the codebase. No to mention a re-write
of settings serialization as you obviously can't write locks to the disk.

Overall I find this to be an acceptable compromise.

Price is the only thing we must update at runtime so it is refreshed from the
main settings object every 10 minutes. This will block one of the many worker
threads while it waits for all other worker threads to release their read locks,
updates the value and moves on. This only requires a single read lock on the main
settings and won't block the main thread. Two big advantages.

Remember actual exit billing is synced with the clients through endpoint hitting
so this won't affect payments. But it will create slight discrepencies in the user
local accounting since local billing won't update the updated price change.