Add protection for full node communication
After discovering that nodes would overpay the exit I've been looking
to prevent it from happening again comperhensively.

My first attempt was to refund incoming transactions immediately and
add long timeouts to the transaction checking. While refunding incoming
transactions has worked quite well to unwind our overpayment balances
long timeouts has served it's purpose but caused other problems.

Like a generally huge number of timeouts in the common loop.

This patch is a replacement for the long timeouts patch where instead
of making very long lived futures we simply keep track of if we have
actually talked to a full node about any given transaction. If we
have not succeeded we panic.

Both servers and clients have an auto restart catch. In some cases
this won't help, in others where the node is merely overloaded it will

In the specific case of the exit running slowly and not validating anything
this would resolve the issue nicely.