Correct payment validator logic to prevent double payments
Previously we used independent do_send statements in order to validate
payments. This is logically incorrect as it's possible for the payment
to be accepted in debt keeper before it is removed from the validation list
in payment validator.

This patch corrects that error by making sure that in order to send debt
keeper messages the removal from the existing list must first succeed.

The next race conditon to consider is that multiple calls to remove
may be running in parallel. We address this  by having remove return
an error when no such transaction exists. Since Actors themselves
serialize access to self this means removal from the list is fully
atomic.