More logical config management for port toggling
As noted in my previous commit there's a hole in this logic where
mesh ports will not be removed from the peer interfaces list when
they are toggled to WAN or LAN. This didn't have negative effects
during testing because the HashSet did not allow duplicate inserts
and would silently allow the entry to be 'added back' later without
issues. And of course having Rita listen on the now WAN or LAN port
didn't cause issues becuase they where not being used that way anyways.

The source of this problem was that config state management was handled
elsewhere from the main logic. This is a poor way to achive the rollback
feature we wanted.

This patch fixes the rollback feature to backup and restore the entire
network settings struct and also cleans up the logic as well.