Applying incoming credit immediately should be configurable
While I was correct in that this wouldn't be an issue for clients I underestimated
how disruptive it would be to balance checking and other debugging. If this where
to remain on network wide routers would end up in a correct non-enforcing state yes
but with no balance convergence.

As routers reboot and forgive debts in memory and remember overpayment they would end
up overpaying each other in a cycle of generosity that's stable (would not result in
more money being sent) but would result in incomprehensible balances until one side
or the other forced convergence by using up all the credit.

Since having to use up all the credit was how things behaved before, we can just do
things that way, have the same behavior, and keep our ability to easily see if
billing agrees or disagrees instantly.