Add get_hash() and get_hash_array() to identity
This lets us grab the hash of the ID easily. For code that doesn't
like to use the hash trait.

Note the use of to_be_bytes in the array function to ensure that
if we for some reason decide to send this hash value over the
wire it will match on little and big endian systems