Encrypt Heartbeat
This contains enough info that I would really feel better if it was encrypted.

I've hardcoded the key for now, I'm not super happy with that but I'm also frustrated
with all the state management being done in SETTINGS right now as a sort of lazy cop out.

We really need to figure out a better management system for the various levels of storage
balance gets changed every 5 seconds, our private key gets written to once. They really shouldn't
be under the same lock with the same level of mutability.

Anyways that's a problem for another time. This commit also drops the encrypted heartbeat
as serde's serialization is too inefficient to fit into a single udp packet. Instead we
byte pack ourselves.