Fix no free tier with phone clients.
The problem that inspired this rabbit hole of a fix was that
when the free tier was disabled for the router users phone
client connectivity would break. It turns out this was a symptom
of a larger misunderstanding about forward rules.

We don't really need any forward rules, our default firewall settings
take care of that. So those got removed and commented in this commit.

Removing the entire masquerade rule was correct pre-phone client
but is now far too broad. The best way to block it is now to preempt
the packets in the lan forwarding chain created by openwrt.

This all really highlights that we need to take better control of the
system iptables or at least have a more detailed understanding of how
firewall3 interacts with our own stuff.

There was also an edge case where toggling client can use freetier
from the settings endpoint would result in incorrect behavior due
to an overly restrictive match statement.