Forbidden settings merge values
We're not working to give network operators more direct access to the oracle
features. So it's important that it be a safe environment. While it's impossible
for us to read the eth_private_key remotely from any router the settings merge json
does provide a chance to overwrite it.

This really isn't acceptible froma saftey perspective. This patch implements a small
but effective blacklist of settings values where local router state is stored in such
a way that remote updates can't possibly be anything but destructive.