Use liboping for ping actions in althea_kernel_interface
This change pulls in the c library oping and a Rust crate providing
bindings.

This lets us perform ping operations without shelling out and most
importantly control the ping timeout with a resolution higher than
one second.

The one second timeout minimum is a key reason why the exits endpoint
can run slowly. As it serially pings all the exits and the inside
exit tunnel ip for the registered exit.

That's at most (Num Exits + 1) * Timeout seconds of waiting time. This
is actually a pretty rare case that all exits are inaccessible AND
we have routes to them in Babel. But sadly this rare case is most commonly
when people are looking at the dashboard! Strange connection troubles, bloat
etc etc. So it has an outsized impact.

This change uses our new found freedom of the timeout to reduce the exit ping
timeout to 200ms. Which is pretty conservative and helps with the maximum wait
time.

Perf Check:
previous common case: 1s
previous worst case: 10s
new common case: 900ms
new worst case: 1.6s

So ultimately
+1 for less shelling out
-1 for more C code
+1 for more performance