Pin cross and upgrade mockito
Required to get our tests working again. It seems that the cargo
cross docker image is now using a version of openssl that is too new
for openssl-sys the crate to pick up and successfully cross compile.

So for the time being we will pin to an older version of the container

Mockito seems like a pretty standard rust feature deprication.