Enum for ForwardingProtocolMessage
this is a huge refactor for forwarding protocol message, the more
I worked with the previous strategy the more I found it difficult
to deal with. It was hard to enumerate all message types, hard to match
against them and very hard to deal with parsing arbitrary messages from
the buffer. This new structure dramatically reduces the difficulty of all
three.