Better logging for bridge ticks
this will cause some log spam but for some reason we keep losing the
bridge on specific nodes usually after a few weeks of uptime. I assume
it's some one in a million seqence of failures that puts the bridge
into a state it can't rescue itself from. So lets try and gain some
more insight the next time it happens