Improve Ip Increment logic
This modifies the ip increment logic in the utils file to use the
integrated rust functions for byte conversion to basic integer types.

This results in a much simpiler algorithm across the board and improvements
in correctness and functionality. For example there is no more limitation to
use a factor of 8 as a subnet size.