Fix light client ip management
The light client ip assignent and return system was broken
in several ways. First and foremost LocalIdentities instead of Identities
where used to identify assigned addresses, meaning that a new address
was always assigned each time a request was made.

Tunnel opening was correct and only one tunnel existed for each phone,
this masked the bug by having the ip return process successfully reclaim
all the assigned ip's except for the one actually in use.

But in the case that there is more than one phone client connected the address
garbage collection had it's own bug which prevented it from removing any
more addresses once it found one in use.

The last bug is that address incrementing function never actually failed even
if you exceeded the allowed subnet if that subnet was a /24 instead of a /16 as
it was in this case. This was fixed in the previous patch

And this is how you get 400 assigned light client ip's and non-working
additional light clients.

This patch fixes all of these issues.