The Big Operator Refactor
This is a pretty major refactor to remove the SubnetDAO concept
and fully replace it with the operator and operator server concept.

Major changes are as follows

'Oracle' split into two modules BlockchainOracle and OperatorUpdate
one that handles updating blockchain info and one that handles updating
info from a management backend.

That management backend has also been changed, instead of being a flatfile
on a server it is now a request to the operator tools server that involves
the router indentifying itself. Allowing you to respond with different settings
on a per router basis and do away with the mess of flatfiles that the previous
system was becoming.

The operator settings module was added and a migration was written from the
existing subnet dao structure into this. There are no remaining uses of the
DAO settings struct except where it is migrated. The new operator struct
doesn't have any concept of multiple operator addresses which drastically
simplifies what was previously pretty hacky (picking the last set DAO address
for the oracle).

The endpoints that the dashboard uses to interact with the DAO have been
maintained for backwards compatiblity but no longer interact with the DAO
settings object and will function to create a single entry list.

The exits where dealing with having a pricing oracle they had
to disable and a dao module they never used, now both of those are moved into
Rita client.

Changing blockchains from the operator update was improved to actually change
the net version and other required info as well.