Add Exit loop watchdog thread and grab data from debt keeper syncronously
The goal of moving Rita exit loop into it's own thread is to prototype a removal
of futures from key repeatedly run logic that is not operated via http endpoint.

async works great for tcp streams, not to great for regularly scheduled tasks
that are absoutely critical for contiued system function.

To that end I want to move all of our loops into their own threads and then slowly
constrain futures to where they make sense. Eg the http endpoints this system
requires. This will require converting all of our 'actors' into files with thread
safe global state and a series of associated function calls.

This is another incremental step along that path, as we've both improved the loop
with a self restart and also added a syncronous way to access the debts from debt
keeper.