Database requests should fail properly
Had a database downtime incident that managed to take down everything
becuase the exit was still up and making peer tunnels but the database
could not be accessed. To further complicate matters there was no recovery
when the database game back because the system was too clogged when time
delayed database requests to actually serve any requests.