The exit should not start successfully unless it can setup clients
The way we handle failover is at the babel level, if there is an exit
up making connections with peers then devices will route to it. This
code causes an exit to fail if it doesn't already have the wg exit
tunnel setup. This prevents exits from running, taking client traffic
but not being able to route it to the outside world