Get exit debt using the sync call
I'm seeing a lot of debts request failures on the exit right now,
this should be faster as many threads can read this object at the
same time. I do have some concerns about locking it and how long that
will bock DebtKeeper during it's update but this will no doubt be faster
in general for the endpoints. Not to mention the complexity advantage