Babel should wait up to 500ms for reads to finish
While our up to 5ms wait works fine at small scales it causes
problems on larger nodes. In order to strike the balance between
waiting for as little time as possible and waiting long enough
for babel to finish we now wait in 10ms increments for up to half
a second.