Adjust Cake flows in place, use Internet bandwidth preset
Two small changes to our traffic shaping flow. First off instead of
deleting and re-creating the flow shaper every time we want to change
it we simply apply the delta using the 'change' parameter.

second we should be using 'internet' not 'metro'. The value of interval
should be set to approximate the round trip time of the stream to it's
final destination out on the internet, not to the exit or the next hop.

Therefore we shouldn't tune this so tightly unless the exit is very far
away.