Polish operator fee payment
Looking at a well placed high-uptime node we observe nearly 20%
operator underpayment. While I don't think these changes make
up all of that they will at least get somewhere.

First off instead of sending the payment threshold we send the
actual amount we owe once it is larger than the threshold. In addition
we capture the sent instant before trying to send the transaction
so that the time that the transaction is being transmitted does not
go unbilled.