Reset network info stats on a timer
The way we where computing our network latency averages advantages
connections that started stable and had a longer time to build up
an average history. By having a longer history their average was
less likley to meet the criteria and less likley to be reset in a
self feeding cycle.

This probably wasn't so bad because links tend to remain stable
but it's better to have everything on a level playing field.

This new system has it's own inherent advantages, links that are changed
more often will gather more and more history rather than having reset
triggered. Hopefully this may counterbalance the desire to constantly
change links that are very unstable