Shape interfaces in bulk
The new system seems to be sending too many IncreaseSpeed messages
resulting in a situation where they are often drowned out as the
message buffer becomes full on the other end. In order to ensure
reliable delivery we will be sending larger messages with a batch
of interfaces to shape.

This should keep us down to a single message per loop iteration.

Also modified here is additional loggic for the 'shaping up at limit'
case and a reset of the last changed param when we attempt shaping
which I thought I had included before (my last commit message on this
code makes references to it) but I seem to have forgotten to actually
include.