Use purely std-dev based latency heuristics
Multiplying by the average disadvantaged the connections that
often needed shaping the most (stable connections that have load
induced instability) at the advantage of noisy connections.

Here I'm trying to strike a balance between keeping latency down
and being sensitive to instability and the reality that there are
many perfectly ok connections sdt-dev in the hundreds, mostly because
they are more borderline links.

In general stable connecitons have a very small std-dev and unstable
connections have an enormous one. So lets focus in on this and see
what we can do.