Compare shaper latency to lowest seen latency
So reseting more often has brought us closer to the median, which is good
in that it's more accurate picture of the short term but bad in that
our std-dev metric doesn't work because we're seeing the deviation
of a bad connection and not taking sufficiently harsh action.

I have tried to avoid bringing in the absolute value of the latency
as it's likley to cause problems with far away exits. But I think
this combo will work a little better in both cases.

We now sample the lowest latency we have ever seen on a link, we know
the link is capable of this and then we use that an an anchor to try
and keep that latency in the picture. This may still end up being
too strict for some links but I think it's the right path