Fixup remote access when user password is set
Starting in Beta 13 we ship with mesh remote access enabled for users
using managed images. This obviously comes with some security considerations
so we disable password auth. The problem is that once a password is set using
the router password function there is still no possiblity for the user to ssh
into the router from the LAN. This wouldn't be an issue except that router rescue
depends on functioning SSH from the lan in order to rescue routers.

Running seperate dropbear instances results in very strange output, so you can't
really have password access on the LAN and key only access everywhere else. We also
don't want to ship with only a default password or lock users out of their routers.

This is the best immediate compromise of many bad options. I'll have to do some more
thinking on the subject into the future.