Timeout streams
I think we're eventually running out of sockets becuase we don't
time out streams properly.

I've observed that after hours and hours of continuous operation
clients crash. I think the server is crashing too because we're
running out of file descriptors for all the sockets we're holdingi
open.

This is a pretty simple solution to that.