Timestamp bump all tunnels for a peer at once
this change causes us to bump the timestamp for all tunnels from
a single peer regardless of if we got a message on that particular
ifidx or with the same source ipv4 address.

I don't think this comes with very serious risks, except perhaps that
the tunnel table will grow too large if somehow a node continuously
changes it's fe80 address or ifidx.

There are some risks here, this code isn't touched much and I'm not
totally sure why the ifidx or fe80 address would change in the first
place. It may be for reasons that invalidate this