Drastically increase tunnel timeout
I'm seeing incorrect timouts occuring on the main Clatskine gateway
this is a large node handling nearly 100 tunnels and I think what's
happening is that the Ubnt radios are damping broadcast traffic across
the bridge or the gateway can't handle all the hello checkins and is
therefore closing tunnels it really shouldn't. To combat this I'm
increasing the timeout so that success can be very very rare and things
will still work just fine.