Add vendored OpenSSL feature
This feature is useful mainly because we can get cross tests to work
again without doing too much work. It bundles an openssl build provided
by the crate maintainers. Which is easier and more self contained than using
system openssl at the risk that if the system openssl is updated your rust
binary will not be updated.

This adds about 3mb to the binary size and as such as a non-starter for the
OpenWRT version of Rita. Both because that's a big file size increase and
because we woudl still need the OpenWRT ssl package for things like opkg
update against an ssl server.