Remove target ip during forwarding start
On high uptime nodes there's a 100% chance that a randomly chosen
former ip will eventually become a target ip and need to be cleared
out of the interface list.

Previously we only cleared out old source ip addresses, this is a very
simple patch to also check for and delete addresses that we are currently
targeting that we previously randomly generated as our source.