Refactor contact info storage and usage
This patch moves contact info storage into a totally new type of enum that
enforces more solid type based validation on what can be provided. It also migrates
everything to that new enum and adds endpoints for getting/setting contact
into in a sane and validated way.

This resolves a lot of my fears about unvalidated contact info. Specifically that
we're parsing things at the endpoints but not in transit and that we may have a
lot of invalid contact info lying around everywhere.

finally we remove a few unwraps hanging around from when we had much lower standards
about panicing.