Check last handshake time before removing tunnel
This check changes the tunnel GC rules such that if and only if we
are about to remove a tunnel due to a timeout we shell out and check
if the tunnel is active by looking at the last handshake time. If there
has been a valid handshake in the last 15 minutes we interpret the tunnel
as still active and avoid garbage collecting it.

This lets us reduce the tunnel timeout back to 15 minutes from 6 hours
so that improper tunnels can still be cleaned up in a reasonable amount
of time and the system can recover from tunnel errors.

We also remove the configurability of the tunnel timeout. It was orignally
added as a sort of backup feature but the new design fails to keeping tunnels
and we can hopefully survive the slight optionality reduction.