Fix file descriptor errors by reducing DNS timeout
The default DNS timeout in the Actix system resolver actor
was set to 5 seconds with 2 attempts. Meaning a total of 10
seconds for each DNS attempt. Since we run many requests on
5 second loops this resulted in the production of an ever
increasing backlog of futures if DNS resolution where to suddenly
fail.

This is patched in our own fork of actix which is otherwise identical
to upstream

other changes included here is a general Cargo.lock update