Add ContactStorage
The TOML crate doesn't allow serialization of enums with struct members.
To resolve this I add ContactStorage a serializable type that is
easy to convert back and forth into the ContactType enum which I plan
to continue using everywhere else.

We can phase out ContactStorage once the TOML crate is improved. This
is a few hundred lines of boilerplate conversions and tricks to use
.into all over the place.