Refuse to de-register clients for the time being
After yesterdays incident where we de-registered everyone in both
production deployments I'm going to be a little cautious before
re-enabling this code.

For the time being the very few devices that don't check in for 60
days, get reset, and then somehow come back online will just have
to be confused until the user manually resets. Small price to pay
for piece of mind.