Migrate contact details immediately before exit registration
This patch provides backwards compatibility with old verisons of the
dashboard attempting to register to exits using Beta 14 rita. Because
the old data was set directly by the settings endpoint the dashboard
ends up sneaking the data into the config unmigrated. So running the
migration only on reboot isn't good enough.